Move format into utility/
