+++ include/bitcoin/bitcoin/impl/utility/format.ipp
