+++ include/bitcoin/bitcoin/impl/utility/serializer.ipp
@@ -20,7 +20,9 @@
 #ifndef LIBBITCOIN_SERIALIZER_IPP
 #define LIBBITCOIN_SERIALIZER_IPP
 
+#include <boost/asio/streambuf.hpp>
 #include <bitcoin/bitcoin/utility/assert.hpp>
+#include <bitcoin/bitcoin/utility/format.hpp>
 
 namespace libbitcoin {
