+++ src/math/ec_keys.cpp
@@ -21,7 +21,7 @@
 
 #include <algorithm>
 #include <secp256k1.h>
-#include <bitcoin/bitcoin/format.hpp>
+#include <bitcoin/bitcoin/utility/format.hpp>
 #include <bitcoin/bitcoin/math/hash.hpp>
 #include <bitcoin/bitcoin/utility/assert.hpp>
