+++ src/script.cpp
@@ -23,7 +23,7 @@
 #include <boost/algorithm/string.hpp>
 #include <boost/optional.hpp>
 #include <bitcoin/bitcoin/constants.hpp>
-#include <bitcoin/bitcoin/format.hpp>
+#include <bitcoin/bitcoin/utility/format.hpp>
 #include <bitcoin/bitcoin/primitives.hpp>
 #include <bitcoin/bitcoin/script.hpp>
 #include <bitcoin/bitcoin/transaction.hpp>
